+++ components/FormWidget/AddMAD/Nicovideo/NicovideoOriginalSource.tsx
@@ -6,7 +6,7 @@ import { CoolImage } from "~/components/CoolImage";
 import { ExternalLinkPictogram } from "~/components/Pictogram";
 import { FragmentType, graphql, useFragment } from "~/gql";
 
-import { TagButton, TagButtonFragment } from "./TagButton";
+import { TagButton, TagButtonFragment } from "../TagButton";
 
 export const NicovideoOriginalSourceFragment = graphql(`
   fragment NicovideoForm_OriginalSource on NicovideoOriginalSource {
